+++ packages/snaps-controllers/src/snaps/endowments/enum.ts
@@ -1,6 +1,5 @@
 export enum SnapEndowments {
   NetworkAccess = 'endowment:network-access',
-  LongRunning = 'endowment:long-running',
   TransactionInsight = 'endowment:transaction-insight',
   Cronjob = 'endowment:cronjob',
   EthereumProvider = 'endowment:ethereum-provider',