+++ packages/snaps-execution-environments/src/common/endowments/crypto.test.ts
@@ -32,7 +32,6 @@ describe('Crypto endowment', () => {
     });
 
     Object.assign(globalThis, {
-      ...globalThis,
       crypto: undefined,
       SubtleCrypto: undefined,
     });