+++ packages/snaps-ui/CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0]
+### Changed
+- Initial stable release from main branch ([#1757](https://github.com/MetaMask/snaps/pull/1757))
+
+## [0.37.5-flask.1]
+### Changed
+- Bump `metamask/utils` and `metamask/snaps-registry` ([#1738](https://github.com/MetaMask/snaps/pull/1738), [#1694](https://github.com/MetaMask/snaps/pull/1694))
+
 ## [0.37.4-flask.1]
 ### Fixed
 - Remove unused dependencies ([#1680](https://github.com/MetaMask/snaps/pull/1680))
@@ -20,7 +28,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - The version of the package no longer needs to match the version of all other
     MetaMask Snaps packages.
 
-[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/snaps-ui@0.37.4-flask.1...HEAD
+[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/snaps-ui@2.0.0...HEAD
+[2.0.0]: https://github.com/MetaMask/snaps/compare/@metamask/snaps-ui@0.37.5-flask.1...@metamask/snaps-ui@2.0.0
+[0.37.5-flask.1]: https://github.com/MetaMask/snaps/compare/@metamask/snaps-ui@0.37.4-flask.1...@metamask/snaps-ui@0.37.5-flask.1
 [0.37.4-flask.1]: https://github.com/MetaMask/snaps/compare/@metamask/snaps-ui@0.37.3-flask.1...@metamask/snaps-ui@0.37.4-flask.1
 [0.37.3-flask.1]: https://github.com/MetaMask/snaps/compare/@metamask/snaps-ui@0.37.2-flask.1...@metamask/snaps-ui@0.37.3-flask.1
 [0.37.2-flask.1]: https://github.com/MetaMask/snaps/releases/tag/@metamask/snaps-ui@0.37.2-flask.1